Last commit

+++ Tema2/server.js
@@ -1,23 +1,40 @@
 const http = require('http')
-const { getProducts , getProduct, createProduct, updateProduct, deleteProduct} = require('./controllers/productController')
+const { getProducts , getProduct, createProduct, updateProduct, deleteProduct, putProduct} = require('./controllers/productController')
 
 const server = http.createServer((req, res) => {
     if(req.url === '/api/products' && req.method === 'GET'){
+
         getProducts(req, res)
-    } else if(req.url.match(/\/api\/products\/([0-9]+)/) && req.method === 'GET'){
+
+    } else if(req.url.split('/').length===4 && req.url.match(/\/api\/products\/([0-9]+)/) && req.method === 'GET'){
+
         const id = req.url.split('/')[3]
         getProduct(req, res, id)
-    } else if(req.url === '/api/products' && req.method === 'POST'){
+
+    } else if(req.url.split('/').length===3 && req.url === '/api/products' && req.method === 'POST'){
+
         createProduct(req, res)
-    } else if(req.url.match(/\/api\/products\/([0-9]+)/) && req.method === 'PUT'){
+
+    } else if(req.url.split('/').length===4 && req.url.match(/\/api\/products\/([0-9]+)/) && req.method === 'PUT'){
+
         const id = req.url.split('/')[3]
-        updateProduct(req, res, id)
-    }else if(req.url.match(/\/api\/products\/([0-9]+)/) && req.method === 'DELETE'){
+        putProduct(req, res, id)
+
+    }else if(req.url.split('/').length===4 && req.url.match(/\/api\/products\/([0-9]+)/) && req.method === 'DELETE'){
+
         const id = req.url.split('/')[3]
         deleteProduct(req, res, id)
+
+    } else if(req.url.split('/').length===4 && req.url.match(/\/api\/products\/([0-9]+)/) && req.method === 'PATCH'){
+
+        const id = req.url.split('/')[3]
+        updateProduct(req, res, id)
+
     } else {
+
         res.writeHead(405, {'Content-Type': 'application/json'})
         res.end(JSON.stringify({message: 'Method Not Allowed'}))
+
     }
 })
